Now builds and tests pass with gcc and clang under MinGW. Also builds with Cygwin, but tests don't start.

+++ UnitTest/NativeJITUnitTest/UnsignedUnitTest.cpp
@@ -64,7 +64,7 @@ namespace NativeJIT
             uint16_t m_p;
             InnerClass* m_innerPointer;
             InnerClass m_innerEmbedded;
-            long long m_q;
+            int64_t m_q;
         };
 #pragma pack(pop)
